requirements.txt
azure.keyvault
azure.identity
azure.functions
azure.cosmos